Fix divide by 0 bug in this recipe
Added new step 4. to check start == end before the divide by 0 scenario. Switched the order of the multiplication and division when initializing the position variable. Performing the division before the multiplication created scenarios where the quotient resulted in a fraction that was then rounded down to 0, causing the multiplication to return 0 when it shouldn't.