Replacing JApplet AudioClip with AudioSystem
JApplet is deprecated and no longer works in newer JDKs. AudioSystem was the way I found to play sounds in native Java.
 (It only works with .wav, .au, .aif, and a few others; not .mp3 or .flac though)

I also removed using the pause method since that doesn't seem to be needed and prompting the user to ask a question since that will be copied over from magic 8 ball anyway.